"It is divided into components"
