+++ week1/project/ecommerce/src/components/productList.js
@@ -1,0 +1,21 @@
+/** @format */
+
+function Products({filteredProducts}) {
+
+  return (
+    <div>
+      <ul className="products">
+        {filteredProducts.map((product) => {
+          return (
+            <li className="productItem" key={product.id}>
+              <img src={product.image} alt={product.title} />
+              <p>{product.title}</p>
+            </li>
+          );
+        })}
+      </ul>
+    </div>
+  );
+}
+
+export default Products;
